lambdas/requirements-runtime.txt
pydantic>=2.0.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
python-docx>=0.8.11
openpyxl>=3.1.0
replicate>=0.25.0
requests>=2.31.0
